overc-cctl: suppress errors for "cube-ctl add" with minimal systemd container
When systemd exists in the most minimal viable fashion in a container,
not all the directories and files that the overc-cctl tries to modify
actually exist.  This patch adds in some conditional run checks for the
directories and files so the code to modify them only fires when they
are present.

[ Issue: LINPUL8-75 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>