cubeit-installer: fix boot failure when --encrypt and --ima-sign used together
The default IMA rules provides the ability of measuring the boot components
and calculating the aggregate integrity value for attesting. However, this
function conflicts with storage-encryption feature which employs PCR policy
session to retrieve the passphrase in a safe way. If the installer enables
both of them, the default IMA rules will be not used.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>