grub configs: enable automatic certificate provision
During the encryption setup in installer, the current state of UEFI
secure boot is considered into as a critical factor to unseal the
passphrase. Hence, the automatic certificate provision is required for
the bootloader employed by the installer image. With a automatic
certificate provision support in the installer image, the state of
UEFI secure boot would be always turned on and the condition of
successfully unsealing the passphrase is the mandatory enablement of
UEFI secure boot.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>