overc-installer: fix the issue cannot copy bzImage
Since the bzImage usually is a soft link to
an absolute path of kernel image, such as
/boot/bzImage-4.1.33-rt23-pulsar-preempt-rt,
thus it will not find the real file if the
file /boot/bzImage-4.1.33-rt23-pulsar-preempt-rt
doesn't exist. So instead of copy this soft link
file, it's better to copy the real file directly.

Signed-off-by: fli <fupan.li@windriver.com>