cubeit: don't enforcing using chainloader command
A new verification mechanism used to verify the integrity of linux kernel
and initrd is implemented, which requires grub bootloader uses linux
and initrd commands to load kernel and initrd and the verification will
occurs in these commands. Hence, either chainloader or linux+initrd
command can provide the verification from now on.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>