grub cfg: don't explicitly set root prefix
The root prefix would be automatically probed if not explicitly
specified and the probed value would point to the device where grub is
installed and it is usually selected by BIOS boot manager. In most cases,
this default behavior is expected, unless an exception occurs.

The usual behavior should determine the setting of the templates, so
explicitly setting root prefix is a bad idea. In most cases, a hard disk
is attached on the target board and the end user may use a USB thumb drive.
This typical hardware setup will be broke by specifying hd0 if both USB
thumb drive and hard disk are on. Without the explicit setting, the boot
device will be simply determined by the BIOS boot order.

Signed-off-by: Lans Zhang <jia.zhang@windriver.com>