overc-installer: fix the issue of run autodev failed with IMA enabled
When IMA feature is enabled, all of the programs should be
signed, otherwise it will be denied when run it. OverC's IMA
signing steps are executed at the installing stage, and the script
autodev missing the IMA signing, thus run the script in a
signed bash instead of run it directly will fix this issue.

Signed-off-by: fupan li <fupan.li@windriver.com>