overc-cctl,cubeit-installer: Remove securityfs from fstab
In the case of running IMA in a container the securitfs will be added
to the container pre-mount setup.  If it is in the fstab generically
without IMA the systemd remount service will fail because securityfs
is not actually mounted.  There is also no point in remounting the
securityfs during the container start in general.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>