overc-cctl: For lxc containers, NOOP any udev 10-dm.rules
Only the root name space can run the 10-dm.rules for udev.  If more
than one entity such as a privileged container invokes the dmsetup
routines to manipulate volumes automatically the root name space will
start unmounting volumes at runtime for the whole system.

This was found when using a special encrypted volume for the system
journal as a shared resource between the containers and the root name
space so as to make the logs all visible in the privileged container.

This type of manipulation is generally only done when the system is
started and it allows the containers to freely start and stop as well
as have the possibility of a r/w /sys.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>