cubeit-installer:  Call partx -a -v to avoid partition mapping failure
If partx was previously used on one of the loop devices for some other
project or reason the partprobe does not always map the partition and
upate the kernel partition table.

If the partx binary is availble call it so that the partition table
will get update properly regardless of who used it last.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>